requirements.txt
asgiref==3.2.10
Django==3.1.2
pytz==2020.1
sqlparse==0.4.1
requests~=2.24.0
paramiko~=2.7.2
dnspython~=2.0.0